# cmd/compile: prioritize non-CALL struct member comparisons