--- a/test/fixedbugs/issue8606.go
+++ b/test/fixedbugs/issue8606.go
@@ -34,6 +34,10 @@ func main() {
 		f any
 		i int
 	}
+	type S4 struct {
+		a [1000]byte
+		b any
+	}
 	b := []byte{1}
 	s1 := S3{func() {}, 0}
 	s2 := S3{func() {}, 1}
@@ -72,6 +76,7 @@ func main() {
 		{false, T3{s: "fooz", j: b}, T3{s: "bar", j: b}},
 		{true, A{s1, s2}, A{s2, s1}},
 		{true, s1, s2},
+		{false, S4{[1000]byte{0}, func() {}}, S4{[1000]byte{1}, func() {}}},
 	} {
 		f := func() {
 			defer func() {